+++ .github/actions/build-deb/entrypoint.sh
@@ -2,7 +2,14 @@
 
 set -e
 
-curl https://sh.rustup.rs -sSf | sh -s -- -y --default-toolchain ${RUST}
+TOOLCHAIN=$(grep -e '^toolchain =' Cargo.toml | sed -e 's/toolchain = "\(.*\)"/\1/')
+
+VERSION=$(grep -e '^version =' Cargo.toml | sed -e 's/version = "\(.*\)"/\1/')
+DEB_VERSION=$(echo "$VERSION" | sed -e 's/-/~/g')
+
+ln -s asm-generic/ /usr/include/asm
+
+curl https://sh.rustup.rs -sSf | sh -s -- -y --default-toolchain ${TOOLCHAIN}
 source $HOME/.cargo/env
 
 rustup target add i686-unknown-linux-gnu
@@ -11,22 +18,18 @@ rustup target add aarch64-unknown-linux-gnu
 
 cargo install cargo-deb
 
-VERSION=$(grep -e '^version =' Cargo.toml | sed -e 's/version = "\(.*\)"/\1/')
-DEB_VERSION=$(echo "$VERSION" | sed -e 's/-/~/g')
-
 mkdir dist
+
+build_deb() {
+  ARCH=$1
+  TARGET=$2
+  cargo deb --target ${TARGET}
+  cp target/${TARGET}/debian/vpncloud_${DEB_VERSION}_${ARCH}.deb dist/vpncloud_${DEB_VERSION}_${ARCH}.deb
+}
 
 cargo deb
 cp target/debian/vpncloud_${DEB_VERSION}_amd64.deb dist/vpncloud_${DEB_VERSION}_amd64.deb
 
-# i386 deb
-cargo deb --target i686-unknown-linux-gnu
-cp target/i686-unknown-linux-gnu/debian/vpncloud_${DEB_VERSION}_i386.deb dist/vpncloud_${DEB_VERSION}_i386.deb
-
-# arm7hf deb
-cargo deb --target armv7-unknown-linux-gnueabihf
-cp target/armv7-unknown-linux-gnueabihf/debian/vpncloud_${DEB_VERSION}_armhf.deb dist/vpncloud_${DEB_VERSION}_armhf.deb
-
-# aarch64 deb
-cargo deb --target aarch64-unknown-linux-gnu
-cp target/aarch64-unknown-linux-gnu/debian/vpncloud_${DEB_VERSION}_arm64.deb dist/vpncloud_${DEB_VERSION}_arm64.deb
+build_deb i386 i686-unknown-linux-gnu
+build_deb armhf armv7-unknown-linux-gnueabihf
+build_deb arm64 aarch64-unknown-linux-gnu