+++ builder/build.sh
@@ -12,9 +12,14 @@ function docker_cmd() {
     vpncloud-builder-$DIST bash -c "$CMD"
 }
 
+# docker run -it --rm -v $(pwd)/..:/home/user/code -v $(pwd)/cache/rpm/target:/home/user/code/target -v $(pwd)/cache/rpm/registry:/home/user/code/registry -v $(pwd)/cache/rpm/git:/home/user/code/git -v $(pwd)/cache/rpm/tmp:/home/user/code/tmp vpncloud-builder-rpm
+
 set -e
 
 cd $(dirname $0)
+
+TOOLCHAIN=$(grep -e '^toolchain =' ../Cargo.toml | sed -e 's/toolchain = "\(.*\)"/\1/')
+UPX_VERSION=$(grep -e '^upx_version =' ../Cargo.toml | sed -e 's/upx_version = "\(.*\)"/\1/')
 
 VERSION=$(grep -e '^version =' ../Cargo.toml | sed -e 's/version = "\(.*\)"/\1/')
 DEB_VERSION=$(echo "$VERSION" | sed -e 's/-/~/g')
@@ -24,12 +29,12 @@ else
   RPM_VERSION="$VERSION-1"
 fi
 
-mkdir -p cache/{rpm,deb}
+mkdir -p cache/{rpm,deb,musl}
 CACHE=$(pwd)/cache
 
 mkdir -p ../dist
 
-docker build --rm -f=Dockerfile-deb -t vpncloud-builder-deb .
+docker build --rm -f=Dockerfile-deb --build-arg TOOLCHAIN=$TOOLCHAIN --build-arg UPX_VERSION=$UPX_VERSION -t vpncloud-builder-deb .
 
 # x86_64 deb
 if ! [ -f ../dist/vpncloud_${DEB_VERSION}_amd64.deb ]; then
@@ -51,19 +56,21 @@ build_deb armhf armv7-unknown-linux-gnueabihf
 build_deb arm64 aarch64-unknown-linux-gnu
 
 
+docker build --rm -f=Dockerfile-musl -t vpncloud-builder-musl .
+
 build_static() {
   ARCH=$1
   TARGET=$2
   if ! [ -f ../dist/vpncloud_${VERSION}_static_${ARCH} ]; then
-    docker_cmd deb "cd code && cargo build --release --features installer --target ${TARGET} && upx --lzma target/${TARGET}/release/vpncloud"
-    cp $CACHE/deb/target/${TARGET}/release/vpncloud ../dist/vpncloud_${VERSION}_static_${ARCH}
+    docker_cmd musl "cd code && cargo build --release --features installer --target ${TARGET} && upx --lzma target/${TARGET}/release/vpncloud"
+    cp $CACHE/musl/target/${TARGET}/release/vpncloud ../dist/vpncloud_${VERSION}_static_${ARCH}
   fi
 }
 
 build_static amd64 x86_64-unknown-linux-musl
-build_static i386 i686-unknown-linux-gnu
+#build_static i386 i686-unknown-linux-musl
 build_static armhf armv7-unknown-linux-musleabihf
-#build_static arm64 aarch64-unknown-linux-musl # fails for unknown reason
+build_static arm64 aarch64-unknown-linux-musl
 
 
 docker build --rm -f=Dockerfile-rpm -t vpncloud-builder-rpm .
@@ -73,3 +80,17 @@ if ! [ -f ../dist/vpncloud_${RPM_VERSION}.x86_64.rpm ]; then
   docker_cmd rpm 'cd code && cargo rpm build'
   cp $CACHE/rpm/target/release/rpmbuild/RPMS/x86_64/vpncloud-${RPM_VERSION}.x86_64.rpm ../dist/vpncloud_${RPM_VERSION}.x86_64.rpm
 fi
+
+build_rpm() {
+  ARCH=$1
+  TARGET=$2
+  if ! [ -f ../dist/vpncloud_${RPM_VERSION}.${ARCH}.rpm ]; then
+    mkdir -p $CACHE/rpm/target
+    [ -L $CACHE/rpm/target/assets ] || ln -s ../assets $CACHE/rpm/target/assets
+    [ -L $CACHE/rpm/target/target ] || ln -s ../target $CACHE/rpm/target/target
+    docker_cmd rpm "cd code && cargo rpm build --target ${TARGET}"
+    cp $CACHE/rpm/target/${TARGET}/release/rpmbuild/RPMS/${ARCH}/vpncloud-${RPM_VERSION}.${ARCH}.rpm ../dist/vpncloud_${RPM_VERSION}.${ARCH}.rpm
+  fi
+}
+
+build_rpm i686 i686-unknown-linux-gnu