+++ src/installer.rs
@@ -1,15 +1,22 @@
-use crate::error::Error;
+use crate::{error::Error, util::run_cmd};
 use std::{
     env,
     fs::{self, File},
     io::Write,
-    os::unix::fs::PermissionsExt
+    os::unix::fs::PermissionsExt,
+    process::Command
 };
 
 const MANPAGE: &[u8] = include_bytes!("../target/vpncloud.1.gz");
 const SERVICE_FILE: &[u8] = include_bytes!("../assets/vpncloud@.service");
 const WS_PROXY_SERVICE_FILE: &[u8] = include_bytes!("../assets/vpncloud-wsproxy.service");
 const EXAMPLE_CONFIG: &[u8] = include_bytes!("../assets/example.net.disabled");
+
+fn systemctl_daemon_reload() {
+    let mut cmd = Command::new("systemctl");
+    cmd.arg("daemon-reload");
+    run_cmd(cmd);
+}
 
 pub fn install() -> Result<(), Error> {
     env::current_exe()
@@ -18,7 +25,7 @@ pub fn install() -> Result<(), Error> {
     fs::set_permissions("/usr/bin/vpncloud", fs::Permissions::from_mode(755))
         .map_err(|e| Error::FileIo("Failed to set permissions for binary", e))?;
     fs::create_dir_all("/etc/vpncloud").map_err(|e| Error::FileIo("Failed to create config folder", e))?;
-    fs::set_permissions("/etc/vpncloud", fs::Permissions::from_mode(600))
+    fs::set_permissions("/etc/vpncloud", fs::Permissions::from_mode(700))
         .map_err(|e| Error::FileIo("Failed to set permissions for config folder", e))?;
     File::create("/etc/vpncloud/example.net.disabled")
         .and_then(|mut f| f.write_all(EXAMPLE_CONFIG))
@@ -32,6 +39,7 @@ pub fn install() -> Result<(), Error> {
     File::create("/lib/systemd/system/vpncloud-wsproxy.service")
         .and_then(|mut f| f.write_all(WS_PROXY_SERVICE_FILE))
         .map_err(|e| Error::FileIo("Failed to create wsporxy service file", e))?;
+    systemctl_daemon_reload();
     info!("Install successful");
     Ok(())
 }
@@ -44,6 +52,7 @@ pub fn uninstall() -> Result<(), Error> {
     fs::remove_file("/lib/systemd/system/vpncloud-wsproxy.service")
         .map_err(|e| Error::FileIo("Failed to remove wsproxy service file", e))?;
     fs::remove_file("/usr/bin/vpncloud").map_err(|e| Error::FileIo("Failed to remove binary", e))?;
+    systemctl_daemon_reload();
     info!("Uninstall successful");
     Ok(())
 }